feat: add cost metrics to OpenAI requests resource
The `CostMetrics` class has been added to calculate the sum of the `cost` field in the `OpenAIRequest` model by days. The result is displayed as a trend with a prefix of `$`. The available ranges for the metric are 30, 60, and 90 days. The results of the metric are cached for 1 minute. The URI key for the metric is set to 'cost-metrics'.

In the `OpenAIRequests` resource, a new field `cost` has been added to display the cost of each request. The field is sortable and displayed as a formatted number with a prefix of `$`. The `CostMetrics` card has been added to the `cards` method to display the cost trend.

In the `CreateChat` class, the `cost` field is set in the `handleResponse` method using the `calculate` method of the `Pricing` class. The `usage